Rename to closer resemble protocol naming
