+++ pkgs/sass_language_services/lib/src/features/document_links/document_link_visitor.dart
@@ -4,7 +4,7 @@ import 'stylesheet_document_link.dart';
 
 typedef UnresolvedLinkData = (StylesheetDocumentLink, bool);
 
-class LinkVisitor with sass.RecursiveStatementVisitor {
+class DocumentLinkVisitor with sass.RecursiveStatementVisitor {
   List<UnresolvedLinkData> unresolvedLinks = [];
 
   @override
