+++ pkgs/sass_language_services/lib/src/language_services_cache.dart
@@ -1,7 +1,7 @@
 import 'package:sass_api/sass_api.dart' as sass;
 import 'package:sass_language_services/sass_language_services.dart';
 
-import 'features/links/stylesheet_document_link.dart';
+import 'features/document_links/stylesheet_document_link.dart';
 
 class CacheEntry {
   TextDocument document;
